+++ corporate/tests/test_remote_billing.py
@@ -195,7 +195,7 @@ class SelfHostedBillingEndpointBasicTest(RemoteRealmBillingTestCase):
             "/self-hosted-billing/not-configured/",
         ]:
             result = self.client_get(url)
-            self.assert_json_error(result, "Must be an organization owner")
+            self.assert_json_error(result, "Must be an organization owner", status_code=403)
 
         # Login as an organization owner to gain access.
         self.login("desdemona")