+++ corporate/tests/test_stripe.py
@@ -5221,6 +5221,7 @@ class RequiresBillingAccessTest(StripeTestCase):
             url: str,
             method: str,
             data: dict[str, Any],
+            status_code: int = 400,
         ) -> None:
             tested_endpoints.add(url)
             for user in users:
@@ -5236,7 +5237,7 @@ class RequiresBillingAccessTest(StripeTestCase):
                     data,
                     content_type="application/json",
                 )
-                self.assert_json_error_contains(result, error_message)
+                self.assert_json_error_contains(result, error_message, status_code)
 
         check_users_cant_access(
             [guest],
@@ -5244,6 +5245,7 @@ class RequiresBillingAccessTest(StripeTestCase):
             "/json/billing/upgrade",
             "POST",
             {},
+            403,
         )
 
         check_users_cant_access(
@@ -5252,6 +5254,7 @@ class RequiresBillingAccessTest(StripeTestCase):
             "/json/billing/sponsorship",
             "POST",
             {},
+            403,
         )
 
         check_users_cant_access(
@@ -5276,6 +5279,7 @@ class RequiresBillingAccessTest(StripeTestCase):
             "/json/upgrade/session/start_card_update_session",
             "POST",
             {},
+            403,
         )
 
         check_users_cant_access(
@@ -5284,6 +5288,7 @@ class RequiresBillingAccessTest(StripeTestCase):
             "/json/billing/event/status",
             "GET",
             {},
+            403,
         )
 
         # Make sure that we are testing all the JSON endpoints