+++ zerver/tests/test_realm.py
@@ -281,7 +281,7 @@ class RealmTest(ZulipTestCase):
 
         self.login("iago")
         result = self.client_patch("/json/realm", data)
-        self.assert_json_error(result, "Must be an organization owner")
+        self.assert_json_error(result, "Must be an organization owner", status_code=403)
 
         self.login("desdemona")
         result = self.client_patch("/json/realm", data)
@@ -338,7 +338,7 @@ class RealmTest(ZulipTestCase):
 
         req = dict(name=new_name)
         result = self.client_patch("/json/realm", req)
-        self.assert_json_error(result, "Must be an organization administrator")
+        self.assert_json_error(result, "Must be an organization administrator", status_code=403)
 
     def test_unauthorized_name_change(self) -> None:
         data = {"full_name": "Sir Hamlet"}
@@ -987,7 +987,7 @@ class RealmTest(ZulipTestCase):
         self.assertFalse(realm.deactivated)
 
         result = self.client_post("/json/realm/deactivate")
-        self.assert_json_error(result, "Must be an organization owner")
+        self.assert_json_error(result, "Must be an organization owner", status_code=403)
         realm = get_realm("zulip")
         self.assertFalse(realm.deactivated)
 
@@ -1632,7 +1632,7 @@ class RealmTest(ZulipTestCase):
 
         req = dict(message_retention_days=orjson.dumps(10).decode())
         result = self.client_patch("/json/realm", req)
-        self.assert_json_error(result, "Must be an organization owner")
+        self.assert_json_error(result, "Must be an organization owner", status_code=403)
 
         self.login("desdemona")
 
@@ -2650,7 +2650,7 @@ class RealmAPITest(ZulipTestCase):
 
         self.login("iago")
         result = self.client_patch("/json/realm", {setting_name: value})
-        self.assert_json_error(result, "Must be an organization owner")
+        self.assert_json_error(result, "Must be an organization owner", status_code=403)
 
         self.login("desdemona")
         result = self.client_patch("/json/realm", {setting_name: value})
@@ -2674,7 +2674,7 @@ class RealmAPITest(ZulipTestCase):
         result = self.client_patch(
             "/json/realm", {setting_name: orjson.dumps({"new": admins_group.id}).decode()}
         )
-        self.assert_json_error(result, "Must be an organization owner")
+        self.assert_json_error(result, "Must be an organization owner", status_code=403)
 
         self.login("desdemona")
         result = self.client_patch(