+++ zerver/views/invite.py
@@ -18,7 +18,12 @@ from zerver.actions.invites import (
     do_send_user_invite_email,
 )
 from zerver.decorator import require_member_or_admin
-from zerver.lib.exceptions import InvitationError, JsonableError, OrganizationOwnerRequiredError
+from zerver.lib.exceptions import (
+    InvitationError,
+    JsonableError,
+    OrganizationAdministratorRequiredError,
+    OrganizationOwnerRequiredError,
+)
 from zerver.lib.response import json_success
 from zerver.lib.streams import access_stream_by_id, get_streams_to_which_user_cannot_add_subscribers
 from zerver.lib.typed_endpoint import ApiParamConfig, PathOnly, typed_endpoint
@@ -43,7 +48,7 @@ def check_role_based_permissions(
         raise OrganizationOwnerRequiredError
 
     if require_admin and not user_profile.is_realm_admin:
-        raise JsonableError(_("Must be an organization administrator"))
+        raise OrganizationAdministratorRequiredError
 
 
 def access_invite_by_id(user_profile: UserProfile, invite_id: int) -> PreregistrationUser: