+++ api_docs/changelog.md
@@ -20,6 +20,19 @@ format used by the Zulip server that they are interacting with.
 
 ## Changes in Zulip 11.0
 
+**Feature level 393**
+
+* `POST /invites`, `POST /invites/{invite_id}/resend`,
+  `DELETE /invites/{invite_id}`: These endpoints now return
+  `code: UNAUTHORIZED_PRINCIPAL` in the error response if the request
+  requires admin privileges and the requesting user is not an
+  organization administrator. Previously, `code: BAD_REQUEST` was
+  returned for these error responses.
+* [Error handling](/api/rest-error-handling): Error responses with
+  `code: UNAUTHORIZED_PRINCIPAL` will now return an HTTP status
+  code of 403. Previously, the HTTP status code for these errors
+  was 400.
+
 **Feature level 392**
 
 * [`GET /users/me/subscriptions`](/api/get-subscriptions),